Initial cut at monitoring the state of the IRC network
This will eventually be used as part of an actual monitoring
system, but in the short term it will simply allow us to do
automated system patching while ensuring that systems don't patch
themselves unless the network is fully healthy.

This includes the addition of a tfes.org package repo, with our
patched version of sic that has TLS support. This patch will be
upstreamed once it's seen a bit of testing on our servers.

I'm also adding the M:Tier package repo so we get -stable package
updates, and rolling out my standard /etc/profile tweaks.