Set up relayd to proxy websocket connections to ngircd
This requires a patched version of relayd to allow switching an
HTTP relay into a TCP one based in a rule. Our patch is publicly
available:

  https://assets.tfes.org/openbsd/6.4/bin/relayd-64tfes1.patch

This will be sent upstream once it's seen a bit of time in
production and I'm confident it doesn't break shit.

With this commit, IRC is now available over a websocket connection
at:

  wss://irc.tfes.org/ws/

This will be used in an exciting new webchat client, coming soon.